Create THERMOCOUPLE
The thermocouple is a temperature sensor that is used to measure the temperature at one specific point and generate output in the form of the EMF or an electric current. The
amount of EMF generated in the thermocouple is very minute (millivolts), this millivolt voltage is to be amplifier using Op-amp.
The amplified op-amp output is taken to the port of the microcontroller from where it is interfaced to ADC module to convert the analog value to digital value. This digitized value has a preset temperature values so the corresponding temperature of the value is communicated to different modules using different protocol such as the temperature is printed in UART terminal on PC and on LCD using SPI communication protocol.
The microcontroller will set the threshold limit for the temperature if it rises it says as a result using buzzer and LED that the temperature has
risen above threshold.